Move the MAC operation structure into the driver structure
Sadly, the Mbed TLS builtin driver for MAC doesn't only rely on Mbed TLS
includes, but uses the PSA hash API for implementing HMAC on top of the
PSA primitives. This triggers a dependency on the MAC operation structure,
which now needs to know how the PSA hash operation structure looks like.

What the final layout should be of these structure headers is TBD.

Signed-off-by: Steven Cooreman <steven.cooreman@silabs.com>